Activation Debug

+++ activate/index.php
@@ -12,15 +12,18 @@
 		$activationId = $row['AccountActivationId'];
 		$userId = $row['UserId'];
 		
-		$stmt = $pdo->prepare('UPDATE Users SET Activated = 1 WHERE Id = :id');
-		$stmt->bindParam(':id', $userId, PDO::PARAM_INT); // <-- Automatically sanitized for SQL by PDO
-		$stmt->execute();
-		
-		$stmt = $pdo->prepare('DELETE FROM AccountActivation WHERE AccountActivationId = :id');
-		$stmt->bindParam(':id', $activationId, PDO::PARAM_INT); // <-- Automatically sanitized for SQL by PDO
-		$stmt->execute();
-		
-		
-		echo 'Activated Account UserId: '.$userId;
+		if($userId != "") {		
+			$stmt = $pdo->prepare('UPDATE Users SET Activated = 1 WHERE Id = :id');
+			$stmt->bindParam(':id', $userId, PDO::PARAM_INT); // <-- Automatically sanitized for SQL by PDO
+			$stmt->execute();
+			
+			$stmt = $pdo->prepare('DELETE FROM AccountActivation WHERE AccountActivationId = :id');
+			$stmt->bindParam(':id', $activationId, PDO::PARAM_INT); // <-- Automatically sanitized for SQL by PDO
+			$stmt->execute();
+			
+			//echo 'Activated Account UserID: '.$userId;
+		}
 	}
+	header("Location: /login/");
+	die();
 ?>
